requirements.txt
qbittorrent-api==2025.11.1
xxhash==3.6.0
python-dotenv==1.2.1
requests==2.32.5
peewee==3.19.0
